requirements.txt
flask
flask-cors
flask_jwt_extended
flask_limiter
passlib
bcrypt<4.0
python-dotenv
Pillow
langchain-google-genai